+++ src/workspace.rs
@@ -171,7 +171,10 @@ impl Workspace {
 
     /// Remove all the contents of all the build directories, freeing disk space.
     pub fn purge_all_build_dirs(&self) -> Result<(), Error> {
-        remove_dir_all(self.builds_dir())?;
+        let dir = self.builds_dir();
+        if dir.exists() {
+            remove_dir_all(&dir)?;
+        }
         Ok(())
     }