+++ CHANGELOG.md
@@ -5,6 +5,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 
+## [0.3.0] - 2019-09-23
+
 ### Added
 
 - New method `Toolchain::rustc` to execute a toolchain's `rustc`.
@@ -68,5 +70,6 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Initial version of Rustwide, extracted from Crater.
 
+[0.3.0]: https://github.com/rust-lang/rustwide/releases/tag/0.3.0
 [0.2.0]: https://github.com/rust-lang/rustwide/releases/tag/0.2.0
 [0.1.0]: https://github.com/rust-lang/rustwide/releases/tag/0.1.0